cleanup et modif type des vars c en long

+++ codeC/station.h
@@ -2,9 +2,9 @@
 
 typedef struct _donneeStation             //raccourcie qui ralonge
 {
-    int ID_station;
-    int consomation;
-    int capacite;
+    unsigned long ID_station;
+    unsigned long consomation;
+    unsigned long capacite;
 } Donnee_station;
 
 void traiterStation (Donnee_station station);
